requirements.txt
fastapi
pydantic
pandas
numpy
boto3
pandapower
networkx
scipy
uvicorn
gunicorn
python-dotenv
tabulate
